created common header component, replaced header with common header component
